monet: colors: Switch to single-precision floating-point math
The only differences in output that I've noticed with Oklab + gamut
clipping are very small differences in some pure blue (#0000FF) colors,
particularly darker ones. The differences are too small to matter, so
let's switch to 32-bit floats for performance.

+++ app/src/main/java/dev/kdrag0n/android12ext/core/xposed/XposedHook.kt
@@ -77,7 +77,7 @@ class XposedHook(
 
             sysuiHooks.applyThemeOverlayController(
                 isGoogle = hasSystemUiGoogle,
-                chromaMultiplier = prefs.getInt("custom_monet_chroma_multiplier", 50).toDouble() / 50,
+                chromaMultiplier = prefs.getInt("custom_monet_chroma_multiplier", 50).toFloat() / 50,
                 multiColor = false,
                 colorOverride = colorOverride,
             )
